Force case sensitivity when matching html tags
Currently the regex match is case insensitive when ignorecase is set. So, with a small vimrc like this:

```
set nocompatible
filetype plugin indent on
set ignorecase
set smartcase
setlocal omnifunc=syntaxcomplete#Complete
```

and an html file like this:

```
<hea
```

When I try and complete `<head>`, the tag is transformed to uppercase `<HEAD>`.

By adding `\C`, the case matching works regardless of `set ignorecase` and returns completion in the expect lowercase form of `<head>`.